+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - Requires PHP `8.4`
 - `Formal\AccessLayer\Connection` is now a final class, all previous implementations are now flagged as internal
 
+### Deprecated
+
+- `Formal\AccessLayer\Query\Select::onDemand()`, use `::lazily()` instead
+- `Formal\AccessLayer\Query\SQL::onDemand()`, use `::lazily()` instead
+
 ### Removed
 
 - `Formal\AccessLayer\Connection\PDO::persistent()`